Address review comment: move some types to ExecuteWorkflowInputs.tsx
