+++ widgets/common/src/ExecuteWorkflowInputs.tsx
@@ -1,11 +1,32 @@
-import type { FunctionComponent } from 'react';
-import type {
-    BaseWorkflowInputs,
-    UserWorkflowInputsState,
-    OnCheckboxChange,
-    OnDropDownChange,
-    OnChange
-} from './executeWorkflow';
+import type { ComponentProps, FunctionComponent } from 'react';
+
+export type OnChange = (event: Event, field: typeof Stage.Basic.UnsafelyTypedFormField) => void;
+export type OnDropDownChange = ComponentProps<typeof Stage.Basic.Dropdown>['onChange'];
+export type OnCheckboxChange = ComponentProps<typeof Stage.Basic.Checkbox>['onChange'];
+
+export type BaseWorkflowInputs = Record<
+    string,
+    {
+        type?: string;
+        default?: string;
+        constraints?: {
+            pattern: string;
+        }[];
+    }
+>;
+
+export type UserWorkflowInputsState = Record<
+    string,
+    | string
+    | number
+    | boolean
+    | string[]
+    | number[]
+    | boolean[]
+    | Record<string, string | number | boolean>
+    | null
+    | undefined
+>;
 
 const t = Stage.Utils.getT('widgets.common.deployments.execute');
